currency entity fixes

+++ src/src/Application/Currency/Command/DeleteCurrencyCommandHandler.php
@@ -5,13 +5,16 @@ declare(strict_types=1);
 namespace App\Application\Currency\Command;
 
 use App\Domain\Currency\Repository\CurrencyRepositoryInterface;
+use App\Domain\Purchase\Resource\PurchaseRepositoryInterface;
 use App\Domain\Shared\Command\CommandHandlerInterface;
+use Symfony\Component\HttpFoundation\File\Exception\AccessDeniedException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class DeleteCurrencyCommandHandler implements CommandHandlerInterface
 {
     public function __construct(
-        private readonly CurrencyRepositoryInterface $currencyRepository
+        private readonly CurrencyRepositoryInterface $currencyRepository,
+        private readonly PurchaseRepositoryInterface $purchaseRepository
     ) {
     }
 
@@ -19,6 +22,10 @@ class DeleteCurrencyCommandHandler implements CommandHandlerInterface
     {
         if (null === $currency = $this->currencyRepository->find($command->id)) {
             throw new NotFoundHttpException('There is no currency with that ID');
+        }
+
+        if(in_array($currency->symbol, $this->purchaseRepository->findUsedCurrencies())){
+            throw new AccessDeniedException("You can not delete currency which is in somebody's wallet");
         }
 
         $this->currencyRepository->remove($currency);
